update lab3 and lab4

+++ lab1/task1/generate_random_arrays.h
@@ -1,11 +1,11 @@
 
 #include <algorithm>
 
-int* generate_random_array(int size, int min, int max)
+long* generate_random_array(long size, long min, long max)
 {
-    int* array = new int[size];
+    long* array = new long[size];
 
-    for (int i = 0; i < size; i++)
+    for (long i = 0; i < size; i++)
     {
         array[i] = rand() % (max - min + 1) + min;
     }
@@ -13,9 +13,9 @@ int* generate_random_array(int size, int min, int max)
     return array;
 }
 
-int* generate_random_sorted_array(int size, int min, int max)
+long* generate_random_sorted_array(long size, long min, long max)
 {
-    int* array = generate_random_array(size, min, max);
+    long* array = generate_random_array(size, min, max);
     std::sort(array, array + size);
 
     return array;
