+++ lab1/task1/simple_search.h
@@ -3,9 +3,9 @@
 
 // simple_search - функция поиска с линейным временем
 // возвращает индекс элемента или -1, если элемента нет.
-int simple_search(int* array, int size, int key)
+long simple_search(long* array, long size, long key)
 {
-    for (int i = 0; i < size; i++)
+    for (long i = 0; i < size; i++)
     {
         if (array[i] == key)
         {
